requirements.dev.txt
pytest~=6.2.4
pytest-cov==2.12.1
pytest-aiohttp~=0.3.0
